requirements.txt
django==4.2.2
dj-database-url==1.2.0
gunicorn==20.1.0
Pillow==9.5.0
Markdown==3.4.4
bleach==6.1.0
whitenoise==6.4.0
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.23.5
requests==2.31.0
opencv-python-headless==4.7.0.72
tensorflow-cpu==2.16.1
protobuf==4.23.1
psycopg2-binary==2.9.9
